requirements.txt
-i https://pypi.org/simple
altair==4.2.2 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.11.2 ; python_full_version >= '3.6.0'
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
bs4==0.0.1
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
entrypoints==0.4 ; python_version >= '3.6'
findspark==2.0.1
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.31 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.17.3 ; python_version >= '3.7'
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
numpy==1.24.2
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3
pillow==9.4.0 ; python_version >= '3.7'
protobuf==3.20.3 ; python_version >= '3.7'
py4j==0.10.9.5
pyarrow==11.0.0 ; python_version >= '3.7'
pydeck==0.8.0 ; python_version >= '3.7'
pygments==2.14.0 ; python_version >= '3.6'
pymongo[srv]==4.3.3
pympler==1.0.1 ; python_version >= '3.6'
pyrsistent==0.19.3 ; python_version >= '3.7'
pyspark==3.2.3
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rich==13.3.2 ; python_full_version >= '3.7.0'
semver==2.13.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
soupsieve==2.4 ; python_version >= '3.7'
streamlit==1.20.0
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.2 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2022.7 ; python_version >= '3.6'
tzlocal==4.2 ; python_version >= '3.6'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
validators==0.20.0 ; python_version >= '3.4'
watchdog==2.3.1 ; platform_system != 'Darwin'
zipp==3.15.0 ; python_version >= '3.7'
